Added README.md file

+++ app/Helpers/DateTimeManipulator.php
@@ -7,7 +7,7 @@ class DateTimeManipulator
 {
     public static function dateDifferenceFromNow($date)
     {
-        $dateReferance = Carbon::createFromDate(date('Y-m-d H:i:s', strtotime($date)));
+        $dateReferance = Carbon::createFromDate((int) date('Y-m-d H:i:s', strtotime($date)));
         $now = now();
         $difference = $dateReferance->diffInYears($now);
